+++ src/main/scala/au/com/agl/arc/load/XMLLoad.scala
@@ -21,7 +21,7 @@ object XMLLoad {
     stageDetail.put("partitionBy", load.partitionBy.asJava)
 
     val saveMode = load.saveMode.getOrElse(SaveMode.Overwrite)
-    stageDetail.put("saveMode", saveMode.toString)     
+    stageDetail.put("saveMode", saveMode.toString.toLowerCase)
 
     val df = spark.table(load.inputView)