+++ src/test/scala/au/com/agl/arc/transform/typing/DoubleTypingSuite.scala
@@ -18,7 +18,7 @@ class DoubleTypingSuite extends FunSuite with BeforeAndAfter {
 
     // Test trimming
     {
-      val col = DoubleColumn(id="1", name="name", description=Some("description"), nullable=true, nullReplacementValue=Some("42.2222"), trim=true, nullableValues="" :: Nil, metadata=None)
+      val col = DoubleColumn(id="1", name="name", description=Some("description"), nullable=true, nullReplacementValue=Some("42.2222"), trim=true, nullableValues="" :: Nil, metadata=None, formatters = None)
 
       // value is null -> nullReplacementValue
       {
@@ -80,7 +80,7 @@ class DoubleTypingSuite extends FunSuite with BeforeAndAfter {
         Typing.typeValue(value, col) match {
           case (res, Some(err)) => {
             assert(res === None)
-            assert(err === TypingError("name", s"Unable to convert '${value}' to double"))
+            assert(err === TypingError("name", s"Unable to convert '${value}' to double using formatters ['#,##0.###;-#,##0.###']"))
           }
           case (_, _) => assert(false)
         }
@@ -90,7 +90,7 @@ class DoubleTypingSuite extends FunSuite with BeforeAndAfter {
 
     // Test null input WITH nullReplacementValue
     {
-      val col = DoubleColumn(id="1", name="name", description=Some("description"), nullable=true, nullReplacementValue=Some("42.2222"), trim=false, nullableValues="" :: Nil, metadata=None)
+      val col = DoubleColumn(id="1", name="name", description=Some("description"), nullable=true, nullReplacementValue=Some("42.2222"), trim=false, nullableValues="" :: Nil, metadata=None, formatters = None)
 
       // value.isNull
       {
@@ -128,7 +128,7 @@ class DoubleTypingSuite extends FunSuite with BeforeAndAfter {
 
     // Test null input WITHOUT nullReplacementValue
     {
-      val col = DoubleColumn(id="1", name="name", description=Some("description"), nullable=false, nullReplacementValue=None, trim=false, nullableValues="" :: Nil, metadata=None)
+      val col = DoubleColumn(id="1", name="name", description=Some("description"), nullable=false, nullReplacementValue=None, trim=false, nullableValues="" :: Nil, metadata=None, formatters = None)
 
       // value.isNull
       {
@@ -166,7 +166,7 @@ class DoubleTypingSuite extends FunSuite with BeforeAndAfter {
     }
     // Test other miscellaneous input types
     {
-      val col = DoubleColumn(id = "1", name = "name", description = Some("description"), nullable = false, nullReplacementValue = None, trim = false, nullableValues = "" :: Nil, metadata=None)
+      val col = DoubleColumn(id = "1", name = "name", description = Some("description"), nullable = false, nullReplacementValue = None, trim = false, nullableValues = "" :: Nil, metadata=None, formatters = None)
 
       // value contains non number/s or characters
       {
@@ -174,7 +174,7 @@ class DoubleTypingSuite extends FunSuite with BeforeAndAfter {
         Typing.typeValue(value, col) match {
           case (res, Some(err)) => {
             assert(res === None)
-            assert(err === TypingError("name", s"Unable to convert '${value}' to double"))
+            assert(err === TypingError("name", s"Unable to convert '${value}' to double using formatters ['#,##0.###;-#,##0.###']"))
           }
           case (_, _) => assert(false)
         }
@@ -187,7 +187,7 @@ class DoubleTypingSuite extends FunSuite with BeforeAndAfter {
         Typing.typeValue(value, col) match {
           case (res, Some(err)) => {
             assert(res === None)
-            assert(err === TypingError("name", s"Unable to convert '${value}' to double"))
+            assert(err === TypingError("name", s"Unable to convert '${value}' to double using formatters ['#,##0.###;-#,##0.###']"))
           }
           case (_, _) => assert(false)
         }
@@ -200,7 +200,7 @@ class DoubleTypingSuite extends FunSuite with BeforeAndAfter {
         Typing.typeValue(value, col) match {
           case (res, Some(err)) => {
             assert(res === None)
-            assert(err === TypingError("name", s"Unable to convert '${value}' to double"))
+            assert(err === TypingError("name", s"Unable to convert '${value}' to double using formatters ['#,##0.###;-#,##0.###']"))
           }
           case (_, _) => assert(false)
         }
@@ -224,11 +224,62 @@ class DoubleTypingSuite extends FunSuite with BeforeAndAfter {
         Typing.typeValue(value, col) match {
           case (res, Some(err)) => {
             assert(res === None)
-            assert(err === TypingError("name", s"Unable to convert '${value}' to double"))
+            assert(err === TypingError("name", s"Unable to convert '${value}' to double using formatters ['#,##0.###;-#,##0.###']"))
           }
           case (_, _) => assert(false)
         }
       }
     }
+
+    //test formatter change negative suffix
+    {
+      val col = DoubleColumn(id = "1", name = "name", description = Some("description"), nullable = false, nullReplacementValue = None, trim = false, nullableValues = "" :: Nil, metadata=None, formatters = Option(List("#,##0.###;#,##0.###-")))
+
+      // value contains negative number
+      {
+        val value = "42.22-"
+        Typing.typeValue(value, col) match {
+          case (Some(res), err) => {
+            assert(res === -42.22)
+            assert(err === None)
+          }
+          case (_, _) => assert(false)
+        }
+      }
+    }  
+
+    //test multiple formatter
+    {
+      val col = DoubleColumn(id = "1", name = "name", description = Some("description"), nullable = false, nullReplacementValue = None, trim = false, nullableValues = "" :: Nil, metadata=None, formatters = Option(List("#,##0.###;#,##0.###-", "#,##0.###;(#,##0.###)")))
+
+      // value contains negative number
+      {
+        val value = "(42.22)"
+        Typing.typeValue(value, col) match {
+          case (Some(res), err) => {
+            assert(res === -42.22)
+            assert(err === None)
+          }
+          case (_, _) => assert(false)
+        }
+      }
+    }  
+
+    //test formatter in error message
+    {
+      val col = DoubleColumn(id = "1", name = "name", description = Some("description"), nullable = false, nullReplacementValue = None, trim = false, nullableValues = "" :: Nil, metadata=None, formatters = Option(List("#,##0.###;#,##0.###-")))
+
+      // value contains negative number
+      {
+        val value = "-42.22"
+        Typing.typeValue(value, col) match {
+          case (res, Some(err)) => {
+            assert(res === None)
+            assert(err === TypingError("name", s"Unable to convert '${value}' to double using formatters ['#,##0.###;#,##0.###-']"))
+          }
+          case (_, _) => assert(false)
+        }
+      }
+    }     
   }
 }