+++ src/test/scala/au/com/agl/arc/util/TestDataUtils.scala
@@ -32,25 +32,30 @@ object TestDataUtils {
         // if both are empty ignore
         if (expected.count != 0 || actual.count != 0) {
 
-            val expectedHashDF = expected.withColumn("_data", to_json(struct(expected.columns.map(col):_*))).withColumn("_hash", sha2(to_json(struct(expected.columns.map(col):_*)),512))
-            val actualHashDF = actual.withColumn("_data", to_json(struct(actual.columns.map(col):_*))).withColumn("_hash", sha2(to_json(struct(actual.columns.map(col):_*)),512))
+            val expectedHashDF = expected.withColumn("_hashLeft", sha2(to_json(struct(expected.columns.sorted.map(col):_*)),512))
+            val actualHashDF = actual.withColumn("_hashRight", sha2(to_json(struct(actual.columns.sorted.map(col):_*)),512))
+
             val transformedDF = expectedHashDF
-                .joinWith(actualHashDF, expectedHashDF("_hash") === actualHashDF("_hash"), "full")
+                .joinWith(actualHashDF, expectedHashDF("_hashLeft") === actualHashDF("_hashRight"), "full")
                 .withColumnRenamed("_1", "expected")
                 .withColumnRenamed("_2", "actual")
 
-            val expectedExceptActual = transformedDF.filter(col("actual").isNull)
-            val actialExceptExpected = transformedDF.filter(col("expected").isNull)
-            val expectedExceptActualCount = expectedExceptActual.count
-            val actialExceptExpectedCount = actialExceptExpected.count     
+            transformedDF.persist 
 
-            if (expectedExceptActualCount != 0 || actialExceptExpectedCount != 0) {
+            val expectedExceptActual = transformedDF.filter(col("actual").isNull)
+            val actualExceptExpected = transformedDF.filter(col("expected").isNull)
+            val expectedExceptActualCount = expectedExceptActual.count
+            val actualExceptExpectedCount = actualExceptExpected.count     
+
+            if (expectedExceptActualCount != 0 || actualExceptExpectedCount != 0) {
                 println("EXPECTED")
                 println(expected.schema)
                 expected.show(false)
                 println("ACTUAL")
                 println(actual.schema)
                 actual.show(false)
+
+                transformedDF.unpersist
                 false
             } else {
                 true